Move rv and sv to v5 commands
