+++ pros/cli/misc_commands.py
@@ -36,35 +36,3 @@ def upgrade(force_check, no_install):
                 ui.logger(__name__).error(f'This manifest cannot perform the upgrade.')
                 return -3
             ui.finalize('upgradeComplete', manager.perform_upgrade())
-
-@misc_commands_cli.command(aliases=['sv'], short_help='Set a kernel variable on a connected V5 device')
-@default_options
-@click.argument('variable', type=click.Choice(['teamnumber', 'robotname']), required=True)
-@click.argument('value', required=True, type=click.STRING, nargs=1)
-def set_variable(variable, value):
-    import pros.serial.devices.vex as vex
-    from pros.serial.ports import DirectPort
-
-    # Get the connected v5 device
-    port = resolve_v5_port(None, 'system')[0]
-    if port == None:
-        return
-    device = vex.V5Device(DirectPort(port))
-    device.kv_write(variable, value)
-    ui.finalize('setVariable', f'{variable} set to {value[:255]}')
-
-@misc_commands_cli.command(aliases=['rv'], short_help='Read a kernel variable from a connected V5 device')
-@default_options
-@click.argument('variable', type=click.Choice(['teamnumber', 'robotname']), required=True)
-def read_variable(variable):
-    import pros.serial.devices.vex as vex
-    from pros.serial.ports import DirectPort
-
-    # Get the connected v5 device
-    port = resolve_v5_port(None, 'system')[0]
-    if port == None:
-        return
-    device = vex.V5Device(DirectPort(port))
-    value = device.kv_read(variable).decode()
-    ui.finalize('readVariable', f'Value of \'{variable}\' : {value}')
-
